pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "echo-mcp-server"
version = "1.0.0"
description = "A modular FastAPI backend with MCP tools integration for AI automation"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Echo MCP Backend Team", email = "team@echo.com"},
]
keywords = ["fastapi", "mcp", "backend", "api", "tidb", "redis", "websocket"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Framework :: FastAPI",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]

dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "python-multipart==0.0.6",
    "starlette==0.27.0",
    "sqlalchemy==2.0.23",
    "pymysql==1.1.0",
    "alembic==1.12.1",
    "cryptography==41.0.7",
    "redis==5.0.1",
    "aioredis==2.0.1",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==4.1.2",
    "pydantic[email]==2.5.0",
    "pydantic-settings==2.1.0",
    "python-dotenv==1.0.0",
    "aiosmtplib==3.0.1",
    "email-validator==2.1.0",
    "websockets==12.0",
    "httpx==0.25.2",
    "requests==2.31.0",
    "pyjwt==2.8.0",
    "python-dateutil==2.8.2",
    "aiofiles==23.2.1",
    "structlog==23.2.0",
    "prometheus-client==0.19.0",
    "typing-extensions==4.8.0",
    "click==8.1.7",
    "rich==13.7.0",
    "pydantic-core==2.14.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "fakeredis>=2.20.1",
    "black>=23.11.0",
    "flake8>=6.1.0",
    "isort>=5.12.0",
    "mypy>=1.7.1",
    "pre-commit>=3.6.0",
    "bandit>=1.7.5",
    "jupyter>=1.0.0",
    "ipython>=8.17.2",
]

prod = [
    "gunicorn>=21.2.0",
    "supervisor>=4.2.5",
]

docs = [
    "mkdocs>=1.5.3",
    "mkdocs-material>=9.4.8",
    "sphinx>=7.2.6",
    "sphinx-rtd-theme>=1.3.0",
]

test = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-xdist>=3.5.0",
    "pytest-benchmark>=4.0.0",
    "pytest-html>=4.1.1",
    "fakeredis>=2.20.1",
    "httpx>=0.25.2",
]

[project.urls]
Homepage = "https://github.com/your-org/fastapi-mcp-backend"
Documentation = "https://fastapi-mcp-backend.readthedocs.io/"
Repository = "https://github.com/your-org/fastapi-mcp-backend.git"
Issues = "https://github.com/your-org/fastapi-mcp-backend/issues"

[project.scripts]
fastapi-mcp-backend = "app.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["tests*", "docs*"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["app"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "redis.*",
    "aioredis.*",
    "pymysql.*",
    "passlib.*",
    "jose.*",
    "fakeredis.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "auth: marks tests related to authentication",
    "database: marks tests related to database operations",
    "redis: marks tests related to Redis operations",
    "websocket: marks tests related to WebSocket functionality",
    "email: marks tests related to email functionality",
]

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/venv/*",
    "*/.venv/*",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.bandit]
exclude_dirs = ["tests", "venv", ".venv"]
skips = ["B101", "B601"]